MAJOR: Django: Initial restructuring.  File contents hasn't changed.  Moved src.graph to Graph.models.
