+++ Graph/views.py
@@ -1,0 +1,3 @@
+from django.shortcuts import render
+
+# View contains the endpoints on the server for the browser to fetch data
