+++ requirements_dev.txt
@@ -1,2 +1,4 @@
 gfapy==1.0.0
 nose==1.3.7
+Django==2.2.1
+sqlparse==0.3.0
